+++ order-service/order-domain/order-application-service/src/main/java/OrderCreatedEventApplicationListener.java
@@ -1,0 +1,17 @@
+import com.food.ordering.system.order.service.domain.event.OrderCreatedEvent;
+import org.springframework.transaction.event.TransactionalEventListener;
+import ports.output.messgae.publisher.OrderCreatedPaymentRequestMessagePublisher;
+
+public class OrderCreatedEventApplicationListener {
+
+    private final OrderCreatedPaymentRequestMessagePublisher orderCreatedPaymentRequestMessagePublisher;
+
+    public OrderCreatedEventApplicationListener(OrderCreatedPaymentRequestMessagePublisher orderCreatedPaymentRequestMessagePublisher) {
+        this.orderCreatedPaymentRequestMessagePublisher = orderCreatedPaymentRequestMessagePublisher;
+    }
+
+    @TransactionalEventListener
+    void process(OrderCreatedEvent orderCreatedEvent) {
+        orderCreatedPaymentRequestMessagePublisher.publish(orderCreatedEvent);
+    }
+}
