+++ order-service/order-domain/order-application-service/src/main/java/ports/output/messgae/publisher/OrderCreatedPaymentRequestMessagePublisher.java
@@ -4,4 +4,5 @@ import com.food.ordering.system.domain.event.publisher.DomainEventPublisher;
 import com.food.ordering.system.order.service.domain.event.OrderCreatedEvent;
 
 public interface OrderCreatedPaymentRequestMessagePublisher extends DomainEventPublisher<OrderCreatedEvent> {
+
 }
